Refactored migrations
