--- alembic/versions/139f03ef51a4_realtydetails_sqaure_filed_nullable_.py
@@ -1,32 +1,0 @@
-"""RealtyDetails 'sqaure' filed  nullable option changed
-
-Revision ID: 139f03ef51a4
-Revises: d0a2d093a902
-Create Date: 2021-04-01 03:51:52.082729
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '139f03ef51a4'
-down_revision = 'd0a2d093a902'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('realty_details', 'square',
-               existing_type=sa.BIGINT(),
-               nullable=True)
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('realty_details', 'square',
-               existing_type=sa.BIGINT(),
-               nullable=False)
-    # ### end Alembic commands ###
