--- alembic/versions/3e7cb8099ffa_changed_constraints.py
@@ -1,30 +1,0 @@
-"""changed constraints
-
-Revision ID: 3e7cb8099ffa
-Revises: 555004976023
-Create Date: 2021-04-22 00:29:58.913632
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '3e7cb8099ffa'
-down_revision = '555004976023'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint('realty_realty_details_id_key', 'realty', type_='unique')
-    op.drop_constraint('realty_details_original_url_key', 'realty_details', type_='unique')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_unique_constraint('realty_details_original_url_key', 'realty_details', ['original_url'])
-    op.create_unique_constraint('realty_realty_details_id_key', 'realty', ['realty_details_id'])
-    # ### end Alembic commands ###
