--- alembic/versions/42b720e765f9_service_repr_attribute_name_changed_for_.py
@@ -1,28 +1,0 @@
-"""service_repr attribute name changed for 4 core tables
-
-Revision ID: 42b720e765f9
-Revises: d357559a3dcb
-Create Date: 2021-04-21 23:13:17.626795
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '42b720e765f9'
-down_revision = 'd357559a3dcb'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
