--- alembic/versions/4c7196bca03b_deleted_square_and_price_fields.py
@@ -1,30 +1,0 @@
-"""Deleted square and price fields
-
-Revision ID: 4c7196bca03b
-Revises: cdd41796aa24
-Create Date: 2021-04-23 04:29:10.161255
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision = '4c7196bca03b'
-down_revision = 'cdd41796aa24'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('realty_details', 'square')
-    op.drop_column('realty_details', 'price')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('realty_details', sa.Column('price', postgresql.DOUBLE_PRECISION(precision=53), autoincrement=False, nullable=False))
-    op.add_column('realty_details', sa.Column('square', sa.BIGINT(), autoincrement=False, nullable=True))
-    # ### end Alembic commands ###
