+++ alembic/versions/513deced8d8c_added_tables.py
@@ -1,0 +1,166 @@
+"""Added tables
+
+Revision ID: 513deced8d8c
+Revises:
+Create Date: 2021-04-29 21:00:34.815629
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '513deced8d8c'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('operation_type',
+    sa.Column('id', sa.BIGINT(), nullable=False),
+    sa.Column('name', sa.VARCHAR(length=255), nullable=False),
+    sa.Column('self_id', sa.BIGINT(), nullable=False),
+    sa.Column('version', sa.TIMESTAMP(), nullable=False),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('self_id', 'version')
+    )
+    op.create_table('realty_details',
+    sa.Column('id', sa.BIGINT(), nullable=False),
+    sa.Column('floor', sa.BIGINT(), nullable=True),
+    sa.Column('floors_number', sa.BIGINT(), nullable=True),
+    sa.Column('square', sa.Float(), nullable=True),
+    sa.Column('price', sa.Float(), nullable=False),
+    sa.Column('published_at', sa.TIMESTAMP(), nullable=False),
+    sa.Column('original_url', sa.VARCHAR(length=255), nullable=False),
+    sa.Column('version', sa.TIMESTAMP(), nullable=True),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('original_url')
+    )
+    op.create_table('realty_type',
+    sa.Column('id', sa.BIGINT(), nullable=False),
+    sa.Column('name', sa.VARCHAR(length=255), nullable=False),
+    sa.Column('self_id', sa.BIGINT(), nullable=False),
+    sa.Column('version', sa.TIMESTAMP(), nullable=False),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('self_id', 'version')
+    )
+    op.create_table('service',
+    sa.Column('id', sa.BIGINT(), nullable=False),
+    sa.Column('name', sa.VARCHAR(length=128), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('state',
+    sa.Column('id', sa.BIGINT(), nullable=False),
+    sa.Column('name', sa.VARCHAR(length=128), nullable=False),
+    sa.Column('self_id', sa.BIGINT(), nullable=False),
+    sa.Column('version', sa.TIMESTAMP(), nullable=False),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('self_id', 'version')
+    )
+    op.create_table('city',
+    sa.Column('id', sa.BIGINT(), nullable=False),
+    sa.Column('name', sa.VARCHAR(length=128), nullable=False),
+    sa.Column('self_id', sa.BIGINT(), nullable=False),
+    sa.Column('version', sa.TIMESTAMP(), nullable=False),
+    sa.Column('state_id', sa.BIGINT(), nullable=False),
+    sa.ForeignKeyConstraint(['state_id'], ['state.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('self_id', 'version')
+    )
+    op.create_table('operation_type_alias',
+    sa.Column('entity_id', sa.BIGINT(), nullable=False),
+    sa.Column('alias', sa.VARCHAR(length=255), nullable=False),
+    sa.ForeignKeyConstraint(['entity_id'], ['operation_type.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('entity_id', 'alias')
+    )
+    op.create_table('operation_type_to_service',
+    sa.Column('entity_id', sa.BIGINT(), nullable=False),
+    sa.Column('service_id', sa.BIGINT(), nullable=False),
+    sa.Column('original_id', sa.VARCHAR(length=255), nullable=False),
+    sa.ForeignKeyConstraint(['entity_id'], ['operation_type.id'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['service_id'], ['service.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('entity_id', 'service_id')
+    )
+    op.create_table('realty_type_alias',
+    sa.Column('entity_id', sa.BIGINT(), nullable=False),
+    sa.Column('alias', sa.VARCHAR(length=255), nullable=False),
+    sa.ForeignKeyConstraint(['entity_id'], ['realty_type.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('entity_id', 'alias')
+    )
+    op.create_table('realty_type_to_service',
+    sa.Column('entity_id', sa.BIGINT(), nullable=False),
+    sa.Column('service_id', sa.BIGINT(), nullable=False),
+    sa.Column('original_id', sa.VARCHAR(length=255), nullable=False),
+    sa.ForeignKeyConstraint(['entity_id'], ['realty_type.id'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['service_id'], ['service.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('entity_id', 'service_id')
+    )
+    op.create_table('state_alias',
+    sa.Column('entity_id', sa.BIGINT(), nullable=False),
+    sa.Column('alias', sa.VARCHAR(length=255), nullable=False),
+    sa.ForeignKeyConstraint(['entity_id'], ['state.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('entity_id', 'alias')
+    )
+    op.create_table('state_to_service',
+    sa.Column('entity_id', sa.BIGINT(), nullable=False),
+    sa.Column('service_id', sa.BIGINT(), nullable=False),
+    sa.Column('original_id', sa.VARCHAR(length=255), nullable=False),
+    sa.ForeignKeyConstraint(['entity_id'], ['state.id'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['service_id'], ['service.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('entity_id', 'service_id')
+    )
+    op.create_table('city_alias',
+    sa.Column('entity_id', sa.BIGINT(), nullable=False),
+    sa.Column('alias', sa.VARCHAR(length=255), nullable=False),
+    sa.ForeignKeyConstraint(['entity_id'], ['city.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('entity_id', 'alias')
+    )
+    op.create_table('city_to_service',
+    sa.Column('entity_id', sa.BIGINT(), nullable=False),
+    sa.Column('service_id', sa.BIGINT(), nullable=False),
+    sa.Column('original_id', sa.VARCHAR(length=255), nullable=False),
+    sa.ForeignKeyConstraint(['entity_id'], ['city.id'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['service_id'], ['service.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('entity_id', 'service_id')
+    )
+    op.create_table('realty',
+    sa.Column('id', sa.BIGINT(), nullable=False),
+    sa.Column('city_id', sa.BIGINT(), nullable=True),
+    sa.Column('state_id', sa.BIGINT(), nullable=False),
+    sa.Column('realty_details_id', sa.BIGINT(), nullable=False),
+    sa.Column('realty_type_id', sa.BIGINT(), nullable=False),
+    sa.Column('operation_type_id', sa.BIGINT(), nullable=True),
+    sa.Column('version', sa.TIMESTAMP(), nullable=True),
+    sa.Column('service_id', sa.BIGINT(), nullable=False),
+    sa.ForeignKeyConstraint(['city_id'], ['city.id'], ondelete='SET NULL'),
+    sa.ForeignKeyConstraint(['operation_type_id'], ['operation_type.id'], ondelete='SET NULL'),
+    sa.ForeignKeyConstraint(['realty_details_id'], ['realty_details.id'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['realty_type_id'], ['realty_type.id'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['service_id'], ['service.id'], ondelete='CASCADE'),
+    sa.ForeignKeyConstraint(['state_id'], ['state.id'], ondelete='CASCADE'),
+    sa.PrimaryKeyConstraint('id'),
+    sa.UniqueConstraint('city_id', 'state_id', 'realty_details_id', 'service_id', 'realty_type_id', 'operation_type_id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('realty')
+    op.drop_table('city_to_service')
+    op.drop_table('city_alias')
+    op.drop_table('state_to_service')
+    op.drop_table('state_alias')
+    op.drop_table('realty_type_to_service')
+    op.drop_table('realty_type_alias')
+    op.drop_table('operation_type_to_service')
+    op.drop_table('operation_type_alias')
+    op.drop_table('city')
+    op.drop_table('state')
+    op.drop_table('service')
+    op.drop_table('realty_type')
+    op.drop_table('realty_details')
+    op.drop_table('operation_type')
+    # ### end Alembic commands ###
