--- alembic/versions/555004976023_added_realty_version.py
@@ -1,30 +1,0 @@
-"""added realty version
-
-Revision ID: 555004976023
-Revises: d2611fc7c48a
-Create Date: 2021-04-12 10:58:46.981098
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '555004976023'
-down_revision = 'f41aa6ee5c2a'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('realty', sa.Column('version', sa.TIMESTAMP(), nullable=True))
-    op.add_column('realty_details', sa.Column('version', sa.TIMESTAMP(), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('realty_details', 'version')
-    op.drop_column('realty', 'version')
-    # ### end Alembic commands ###
