--- alembic/versions/8de4d2189413_unique_constraint_added.py
@@ -1,28 +1,0 @@
-"""Unique constraint added
-
-Revision ID: 8de4d2189413
-Revises: 213559dc2506
-Create Date: 2021-03-31 22:32:33.728606
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '8de4d2189413'
-down_revision = '213559dc2506'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_unique_constraint(None, 'realty', ['city_id', 'state_id', 'realty_details_id', 'realty_type_id', 'operation_type_id'])
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'realty', type_='unique')
-    # ### end Alembic commands ###
