--- alembic/versions/a04cc8045c2d_added_relation_for_realty_with_service.py
@@ -1,34 +1,0 @@
-"""Added relation for realty with service
-
-Revision ID: a04cc8045c2d
-Revises: e27ca8bb00dd
-Create Date: 2021-04-29 10:05:21.742210
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'a04cc8045c2d'
-down_revision = 'e27ca8bb00dd'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('realty', sa.Column('service_id', sa.BIGINT(), nullable=False))
-    op.drop_constraint('realty_city_id_state_id_realty_details_id_realty_type_id_op_key', 'realty', type_='unique')
-    op.create_unique_constraint(None, 'realty', ['city_id', 'state_id', 'realty_details_id', 'service_id', 'realty_type_id', 'operation_type_id'])
-    op.create_foreign_key(None, 'realty', 'service', ['service_id'], ['id'], ondelete='CASCADE')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'realty', type_='foreignkey')
-    op.drop_constraint(None, 'realty', type_='unique')
-    op.create_unique_constraint('realty_city_id_state_id_realty_details_id_realty_type_id_op_key', 'realty', ['city_id', 'state_id', 'realty_details_id', 'realty_type_id', 'operation_type_id'])
-    op.drop_column('realty', 'service_id')
-    # ### end Alembic commands ###
