--- alembic/versions/cdd41796aa24_cascade_delete_added.py
@@ -1,58 +1,0 @@
-"""Cascade delete added
-
-Revision ID: cdd41796aa24
-Revises: ff65b93addb9
-Create Date: 2021-04-22 16:24:22.519825
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'cdd41796aa24'
-down_revision = 'ff65b93addb9'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint('city_alias_city_id_version_fkey', 'city_alias', type_='foreignkey')
-    op.create_foreign_key(None, 'city_alias', 'city', ['city_id', 'version'], ['self_id', 'version'], ondelete='CASCADE')
-    op.drop_constraint('city_to_service_city_id_version_fkey', 'city_to_service', type_='foreignkey')
-    op.create_foreign_key(None, 'city_to_service', 'city', ['city_id', 'version'], ['self_id', 'version'], ondelete='CASCADE')
-    op.drop_constraint('operation_type_alias_operation_type_id_version_fkey', 'operation_type_alias', type_='foreignkey')
-    op.create_foreign_key(None, 'operation_type_alias', 'operation_type', ['operation_type_id', 'version'], ['self_id', 'version'], ondelete='CASCADE')
-    op.drop_constraint('operation_type_to_service_operation_type_id_version_fkey', 'operation_type_to_service', type_='foreignkey')
-    op.create_foreign_key(None, 'operation_type_to_service', 'operation_type', ['operation_type_id', 'version'], ['self_id', 'version'], ondelete='CASCADE')
-    op.drop_constraint('realty_type_alias_realty_type_id_version_fkey', 'realty_type_alias', type_='foreignkey')
-    op.create_foreign_key(None, 'realty_type_alias', 'realty_type', ['realty_type_id', 'version'], ['self_id', 'version'], ondelete='CASCADE')
-    op.drop_constraint('realty_type_to_service_realty_type_id_version_fkey', 'realty_type_to_service', type_='foreignkey')
-    op.create_foreign_key(None, 'realty_type_to_service', 'realty_type', ['realty_type_id', 'version'], ['self_id', 'version'], ondelete='CASCADE')
-    op.drop_constraint('state_alias_state_id_version_fkey', 'state_alias', type_='foreignkey')
-    op.create_foreign_key(None, 'state_alias', 'state', ['state_id', 'version'], ['self_id', 'version'], ondelete='CASCADE')
-    op.drop_constraint('state_to_service_state_id_version_fkey', 'state_to_service', type_='foreignkey')
-    op.create_foreign_key(None, 'state_to_service', 'state', ['state_id', 'version'], ['self_id', 'version'], ondelete='CASCADE')
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'state_to_service', type_='foreignkey')
-    op.create_foreign_key('state_to_service_state_id_version_fkey', 'state_to_service', 'state', ['state_id', 'version'], ['self_id', 'version'], ondelete='RESTRICT')
-    op.drop_constraint(None, 'state_alias', type_='foreignkey')
-    op.create_foreign_key('state_alias_state_id_version_fkey', 'state_alias', 'state', ['state_id', 'version'], ['self_id', 'version'], ondelete='RESTRICT')
-    op.drop_constraint(None, 'realty_type_to_service', type_='foreignkey')
-    op.create_foreign_key('realty_type_to_service_realty_type_id_version_fkey', 'realty_type_to_service', 'realty_type', ['realty_type_id', 'version'], ['self_id', 'version'], ondelete='RESTRICT')
-    op.drop_constraint(None, 'realty_type_alias', type_='foreignkey')
-    op.create_foreign_key('realty_type_alias_realty_type_id_version_fkey', 'realty_type_alias', 'realty_type', ['realty_type_id', 'version'], ['self_id', 'version'], ondelete='RESTRICT')
-    op.drop_constraint(None, 'operation_type_to_service', type_='foreignkey')
-    op.create_foreign_key('operation_type_to_service_operation_type_id_version_fkey', 'operation_type_to_service', 'operation_type', ['operation_type_id', 'version'], ['self_id', 'version'], ondelete='RESTRICT')
-    op.drop_constraint(None, 'operation_type_alias', type_='foreignkey')
-    op.create_foreign_key('operation_type_alias_operation_type_id_version_fkey', 'operation_type_alias', 'operation_type', ['operation_type_id', 'version'], ['self_id', 'version'], ondelete='RESTRICT')
-    op.drop_constraint(None, 'city_to_service', type_='foreignkey')
-    op.create_foreign_key('city_to_service_city_id_version_fkey', 'city_to_service', 'city', ['city_id', 'version'], ['self_id', 'version'], ondelete='RESTRICT')
-    op.drop_constraint(None, 'city_alias', type_='foreignkey')
-    op.create_foreign_key('city_alias_city_id_version_fkey', 'city_alias', 'city', ['city_id', 'version'], ['self_id', 'version'], ondelete='RESTRICT')
-    # ### end Alembic commands ###
