--- alembic/versions/d2611fc7c48a_delete_cascade_in_both_drections.py
@@ -1,28 +1,0 @@
-"""Delete cascade in both drections
-
-Revision ID: d2611fc7c48a
-Revises: 139f03ef51a4
-Create Date: 2021-04-01 04:22:04.755117
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'd2611fc7c48a'
-down_revision = '139f03ef51a4'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
