--- alembic/versions/d357559a3dcb_columns_name_changed_from__alias_to_.py
@@ -1,28 +1,0 @@
-"""Columns name changed from *_alias to aliases
-
-Revision ID: d357559a3dcb
-Revises: 3e95cbc40d5e
-Create Date: 2021-04-21 22:43:01.019406
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'd357559a3dcb'
-down_revision = '3e95cbc40d5e'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
