+++ service_api/static_data/operation_type.csv
@@ -1,5 +1,4 @@
 name,self_id
 Продаж,1
 Довгострокова оредна,2
-Долгострочная аренда,2
 Подобова оренда,3
